Added Makefile and refractored code
