+++ Makefile
@@ -1,0 +1,5 @@
+build:
+	cd ./backend && go mod download && CGO_ENABLED=0 GOOS=linux go build main.go
+
+run: build
+	cd ./backend && brew services start mongodb-community && go run main.go start-server
