+++ backend/app/Models/models.go
@@ -1,7 +1,5 @@
 package models
 
-//import"go.mongodb.org/mongo-driver/bson/primitive"
-
 type Product struct {
 	Name        string    `json:"name" bson:"name"`
 	Description string    `json:"description" bson:"description"`
