--- backend/config/config.env
@@ -1,1 +1,0 @@
-COOKIE_EXPIRE = 7 //to be used for jwt token
