+++ backend/internal/config/config.env
@@ -1,0 +1,1 @@
+COOKIE_EXPIRE = 7 //to be used for jwt token
